Started with the new lexer implementation
